requirements.txt
tweepy==4.14.0
python-dotenv==0.21.0
schedule==1.1.0